Move waitForBalanceUpdate to helpers file
